Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,9 +2,9 @@
 
 
 int main() {
-    int a,b;
+    float a,b;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%f %f %c",&a,&b,&c);
     if (c=='+'){printf("%d",a+b);}
     else if(c=='-'){printf("%d",a-b);}
     else if(c=='*'){printf("%d",a*b);}
